Add linting and formatting JS tooling
Resolves #3839
