+++ data-drivers/.prettierrc.js
@@ -1,0 +1,5 @@
+import duskJsPrettierConfig from "@dusk-network/prettier-config/js/index.js";
+
+export default {
+  ...duskJsPrettierConfig,
+};
