+++ data-drivers/package.json
@@ -1,0 +1,38 @@
+{
+  "author": {
+    "name": "Dusk Network B.V.",
+    "url": "https://dusk.network/",
+    "mail": "info@dusk.network"
+  },
+  "bugs": "https://github.com/dusk-network/rusk/issues",
+  "engines": {
+    "node": "^20.0.0"
+  },
+  "homepage": "https://github.com/dusk-network/rusk/tree/master/data-drivers",
+  "license": "MPL-2.0",
+  "name": "data-drivers",
+  "private": true,
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/dusk-network/rusk.git"
+  },
+  "version": "0.1.0",
+  "type": "module",
+  "scripts": {
+    "checks": "npm run format && npm run lint",
+    "format": "prettier . --check",
+    "format:fix": "prettier . --write",
+    "lint": "eslint .",
+    "lint:fix": "eslint . --fix",
+    "start": "cd ./stake-contract-test && node test.js"
+  },
+  "devDependencies": {
+    "@dusk-network/eslint-config": "4.0.1",
+    "@dusk-network/prettier-config": "1.1.0",
+    "eslint": "9.33.0",
+    "eslint-config-prettier": "10.1.8",
+    "eslint-plugin-import": "2.32.0",
+    "eslint-plugin-import-newlines": "1.4.0",
+    "prettier": "3.6.2"
+  }
+}
