--- data-drivers/stake-contract-test/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "wasm-test-runner",
-  "version": "0.1.0",
-  "type": "module",
-  "scripts": {
-    "start": "node test.js"
-  },
-  "devDependencies": {},
-  "type": "module",
-  "scripts": {
-    "start": "node test.js",
-    "test": "jest"
-  },
-  "devDependencies": {
-    "jest": "^29.0.0"
-  },
-  "jest": {
-    "testEnvironment": "node"
-  }
-}
